Refactor: 이익률 관련 클래스 분리
